pyproject.toml
[project]
name = "dagster-dlt-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dagster",
    "dagster-cloud",
    "dagster-dlt>=0.26.11",
    "dagster-duckdb>=0.26.12",
    "dagster-embedded-elt>=0.26.11",
    "dagster-webserver>=1.10.11",
    "dlt[duckdb]>=1.9.0",
    "fastexcel>=0.13.0",
    "polars>=1.27.1",
    "xmltodict>=0.14.2",
]
